Creation des objets necesaire à la création de request
